Converted execution of data transfer stuff to go through a single pipeline

+++ src/operations/shift.rs
@@ -17,6 +17,7 @@ impl From<u32> for Shift {
     }
 }
 
+#[derive(Copy, Clone)]
 pub enum ShiftType {
     LogicalLeft = 0b00,
     LogicalRight = 0b01,
@@ -37,16 +38,16 @@ impl From<u32> for ShiftType {
     }
 }
 
-pub fn apply_shift(shift_type: ShiftType, shift_amount: u32, base_value: u32) -> u32{
+pub fn apply_shift(shift_type: ShiftType, shift_amount: u8, base_value: u32) -> u8{
     let shifted_value;
 
     match shift_type {
         ShiftType::LogicalLeft => {
-            shifted_value = base_value << shift_amount;
+            shifted_value = base_value << (shift_amount as u32);
             // todo: make sure flags aren't a thing
         }
         ShiftType::LogicalRight => {
-            shifted_value = base_value >> shift_amount;
+            shifted_value = base_value >> (shift_amount as u32);
             // todo: make sure flags aren't a thing
         }
         ShiftType::ArithmeticRight => {
@@ -54,10 +55,10 @@ pub fn apply_shift(shift_type: ShiftType, shift_amount: u32, base_value: u32) ->
             // make sure this isn't truncating
         }
         ShiftType::RotateRight => {
-            shifted_value = base_value.rotate_right(shift_amount);
+            shifted_value = base_value.rotate_right(shift_amount as u32);
         }
         _ => panic!("Shift type fucked up")
     }
 
-    return shifted_value;
+    return shifted_value as u8;
 }
